Uplift movability and mutability, the simple way
